confusion.matrix {SDMTools}	R Documentation

Confusion Matrix

Description

confusion.matrix calculates a confusion matrix.

Note: this method will exclude any missing data

Usage

confusion.matrix(obs, pred, threshold = 0.5)

Arguments

obs
a vector of observed values which must be 0 for absences and 1 for occurrences

pred
a vector of the same length as obs representing the predicted values. Values must be between 0 & 1 prepresenting a likelihood.

threshold
a single threshold value between 0 & 1

Value

Returns a confusion matrix (table) of class 'confusion.matrix' representing counts of true & false presences and absences.

Author(s)

Jeremy VanDerWal jjvanderwal@gmail.com

See Also

auc, Kappa, omission, sensitivity, specificity, prop.correct, accuracy

Examples

#create some data
obs = c(sample(c(0,1),20,replace=TRUE),NA); obs = obs[order(obs)]
pred = runif(length(obs),0,1); pred = pred[order(pred)]

#calculate the confusion matrix
confusion.matrix(obs,pred,threshold=0.5)

[Package SDMTools version 1.1-221 Index]