getXYcoords {SDMTools}	R Documentation

Computes the X and Y Coordinates of the Pixels of a Raster Map

Description

getXYcoords computes the geographical coordinates of the rows and columns of pixels of a raster map of class asc. Code & helpfile were modified from adehabitat package.

Usage

getXYcoords(w)

Arguments

w
an object of class asc.

Value

Returns a list with two components:

x
the x coordinates of the columns of pixels of the map

y
the y coordinates of the rows of pixels of the map

Author(s)

Jeremy VanDerWal jjvanderwal@gmail.com

Examples

tasc = as.asc(matrix(rep(x=1:10, times=1000),nr=100)); print(tasc)
getXYcoords(tasc)

[Package SDMTools version 1.1-221 Index]