legend.gradient {SDMTools}	R Documentation

Legend Gradient

Description

legend.gradient creates and displays a gradient legend on a plot or image file. The place and size of the legend is defined by coordinates, previously identified.

Usage

legend.gradient(pnts, cols = heat.colors(100), limits = c(0, 1),
  title = "Legend", ...)

Arguments

pnts
x and y coordinates of the gradient location in the plot

cols
a set of 2 or more colors used in the image, to create the gradient

limits
to label the min and max values of the gradient in the legend

title
to specify the title of the legend

...
other graphical parameters defined by image() or plot()

Value

nothing is returned, a gradient legend is added to a plot or a image.

Author(s)

Lorena Falconi lorefalconi@gmail.com

Examples

#define a simple binary matrix
tmat = { matrix(c( 0,0,0,1,0,0,1,1,0,1,
                   0,0,1,0,1,0,0,0,0,0,
                   0,1,NA,1,0,1,0,0,0,1,
                   1,0,1,1,1,0,1,0,0,1,
                   0,1,0,1,0,1,0,0,0,1,
                   0,0,1,0,1,0,0,1,1,0,
                   1,0,0,1,0,0,1,0,0,0,
                   0,1,0,0,0,1,0,NA,NA,NA,
                   0,0,1,1,1,0,0,NA,NA,NA,
                   1,1,1,0,0,0,0,NA,NA,NA),nr=10,byrow=TRUE) }

#do the connected component labeling
tasc = ConnCompLabel(tmat)

# Create a color ramp
colormap=c("grey","yellow","yellowgreen","olivedrab1","lightblue4")

#create an image
image(tasc,col=colormap, axes=FALSE, xlab="", ylab="", ann=FALSE)

#points for the gradient legend
pnts = cbind(x =c(0.8,0.9,0.9,0.8), y =c(1.0,1.0,0.8,0.8))

#create the gradient legend
legend.gradient(pnts,colormap,c("Low","High"))

[Package SDMTools version 1.1-221 Index]