Make presence-only list for multispeciesPP

Make presence-only list for multispeciesPP

Usage

make_mpp_list(x, id)

Arguments

x

A data.frame containing covariate and bias values for locations, and an identity (species) column

id

The name of the identity column in tidyselect form. id must be a column name in x.

Value

A named list or data frames.

Details

The package multispeciesPP requires a very annoying named list format for presence-only data. This function takes a table of covariate / bias values and an ID column (i.e. species), and returns a named list where each element is the values corresponding to that identity only. Ugh.

Examples

if (FALSE) { # \dontrun{
make_mpp_list(presence_only_data, species)
} # }