omission {SDMTools}	R Documentation

Measures of Accuracy

Description

Estimates different measures of accurracy given a confusion matrix.

Usage

omission(mat)

sensitivity(mat)

specificity(mat)

prop.correct(mat)

Arguments

mat
a confusion matrix of class 'confusion.matrix' from confusion.matrix

Value

returns single values representing the:

ommission
the ommission rate as a proportion of true occurrences misidentified given the defined threshold value

sensitivity
the sensitivity given the defined threshold value

specificity
the specificity given the defined threshold value

prop.correct
the proportion of the presence and absence records correctly identified given the defined threshold value

Author(s)

Jeremy VanDerWal jjvanderwal@gmail.com

See Also

auc, Kappa, confusion.matrix, accuracy

Examples

#create some data
obs = c(sample(c(0,1),20,replace=TRUE),NA); obs = obs[order(obs)]
pred = runif(length(obs),0,1); pred = pred[order(pred)]

#calculate the confusion matrix
mat = confusion.matrix(obs,pred,threshold=0.5)

#calculate the accuracy measures
omission(mat)
sensitivity(mat)
specificity(mat)
prop.correct(mat)

[Package SDMTools version 1.1-221 Index]