wt.mean {SDMTools}	R Documentation

Weighted mean, variance and standard deviation calculations

Description

wt.mean calculates the mean given a weighting of the values.

wt.var is the unbiased variance of the weighted mean calculation using equations of GNU Scentific Library (http://www.gnu.org/software/gsl/manual/html_node/Weighted-Samples.htmland.

wt.sd is the standard deviation of the weighted mean calculated as the sqrt of wt.var.

Note: NA data is automatically ommitted from analysis.

Usage

wt.mean(x, wt)

wt.var(x, wt)

wt.sd(x, wt)

Arguments

x
is a vector of numerical data.

wt
is a vector of equal length to x representing the weights.)

Value

returns a single value from analysis requested.

Author(s)

Jeremy VanDerWal jjvanderwal@gmail.com

Examples

#define simple data
x = 1:25 # set of numbers
wt = runif(25) #some arbitrary weights

#display means & variances (unweighted and then weighted)
mean(x); wt.mean(x,wt)
var(x); wt.var(x,wt)
sd(x); wt.sd(x,wt)

[Package SDMTools version 1.1-221 Index]